do not delete scheduled jobs. ignore bad job entries (likely left overs from rq=1.13.0.

+++ socialhome/tasks/apps.py
@@ -2,6 +2,8 @@ import sys
 
 import django_rq
 from django.apps import AppConfig
+from rq.exceptions import InvalidJobOperation
+from rq.job import JobStatus
 
 from socialhome.streams.tasks import streams_tasks
 
@@ -19,7 +21,10 @@ class TasksConfig(AppConfig):
 
         # Delete any existing jobs in the scheduler when the app starts up
         for job in scheduler.get_jobs():
-            job.delete()
+            try:
+                if job.get_status() != JobStatus.SCHEDULED: job.delete()
+            except InvalidJobOperation:
+                pass # log something here?
 
         # Queue tasks
         streams_tasks(scheduler)
